Issue #19002: Move TabsTrayViewHolder to browser package
