+++ app/src/main/java/org/mozilla/fenix/tabstray/browser/TabsTrayListViewHolder.kt
@@ -9,7 +9,6 @@ import androidx.core.content.ContextCompat
 import mozilla.components.concept.base.images.ImageLoader
 import mozilla.components.concept.tabstray.Tab
 import org.mozilla.fenix.R
-import org.mozilla.fenix.tabstray.TabsTrayViewHolder
 import org.mozilla.fenix.selection.SelectionHolder
 import org.mozilla.fenix.tabstray.TabsTrayStore
 import kotlin.math.max
